requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
emoji==2.7.0
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
filetype==1.2.0
fire==0.5.0
flatbuffers==23.5.26
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.279
langchain-experimental==0.0.13
langcorn==0.0.16
langsmith==0.0.22
loguru==0.7.2
lxml==4.9.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.1.1
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
rdflib==7.0.0
regex==2023.8.8
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.4.0
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.9.3
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
